debug readK

+++ simulations/readK.py
@@ -15,15 +15,17 @@ k3 = []
 
 for file in files:
     f=open (file, 'r'). readlines()
-    k1 += list(map(float, f[0].rstrip('\n').split())) 
-    k2 += list(map(float, f[1].rstrip('\n').split()))
-    k3 += list(map(float, f[2].rstrip('\n').split())) 
+    if len(f)>3:
+        continue
+    k1.append(list(map(float, f[0].rstrip('\n').split())))
+    k2.append(list(map(float, f[1].rstrip('\n').split())))
+    k3.append(list(map(float, f[2].rstrip('\n').split())))
     if args.rm:
         os.remove(file)
 
-k1 = np.array(k1).reshape(len(files), len(k1)//len(files))
-k2 = np.array(k2).reshape(len(files), len(k2)//len(files))
-k3 = np.array(k3).reshape(len(files), len(k3)//len(files))
+k1 = np.array(k1)
+k2 = np.array(k2)
+k3 = np.array(k3)
 
 np.savetxt("../output/"+args.dir+"/k1.csv", k1, delimiter=",")
 np.savetxt("../output/"+args.dir+"/k2.csv", k2, delimiter=",")
